+++ crates/rust-client/src/sync/state_sync.rs
@@ -7,6 +7,7 @@ use miden_objects::{
     block::{BlockHeader, BlockNumber},
     crypto::merkle::{InOrderIndex, MmrDelta, MmrPeaks, PartialMmr},
     note::{NoteId, NoteTag},
+    transaction::ChainMmr,
 };
 use tracing::info;
 
@@ -92,23 +93,21 @@ impl StateSync {
     /// 7. The MMR is updated with the new peaks and authentication nodes.
     ///
     /// # Arguments
-    /// * `current_partial_mmr` - The current partial MMR.
+    /// * `current_chain_mmr` - The current chain MMR.
     /// * `accounts` - All the headers of tracked accounts.
     /// * `note_tags` - The note tags to be used in the sync state request.
     /// * `unspent_input_notes` - The current state of unspent input notes tracked by the client.
     /// * `unspent_output_notes` - The current state of unspent output notes tracked by the client.
     pub async fn sync_state(
         self,
-        mut current_partial_mmr: PartialMmr,
+        mut current_chain_mmr: ChainMmr,
         accounts: Vec<AccountHeader>,
         note_tags: Vec<NoteTag>,
         unspent_input_notes: Vec<InputNoteRecord>,
         unspent_output_notes: Vec<OutputNoteRecord>,
         mut uncommitted_transactions: Vec<TransactionRecord>,
     ) -> Result<StateSyncUpdate, ClientError> {
-        let block_num = (u32::try_from(current_partial_mmr.num_leaves() - 1)
-            .expect("The number of leaves in the MMR should be greater than 0 and less than 2^32"))
-        .into();
+        let block_num = current_chain_mmr.chain_length().checked_sub(1).unwrap_or_default();
 
         let mut state_sync_update = StateSyncUpdate {
             block_num,
@@ -120,7 +119,7 @@ impl StateSync {
             if !self
                 .sync_state_step(
                     &mut state_sync_update,
-                    &mut current_partial_mmr,
+                    &mut current_chain_mmr,
                     &accounts,
                     &note_tags,
                 )
@@ -173,7 +172,7 @@ impl StateSync {
     async fn sync_state_step(
         &self,
         state_sync_update: &mut StateSyncUpdate,
-        current_partial_mmr: &mut PartialMmr,
+        current_chain_mmr: &mut ChainMmr,
         accounts: &[AccountHeader],
         note_tags: &[NoteTag],
     ) -> Result<bool, ClientError> {
@@ -216,14 +215,19 @@ impl StateSync {
         let (new_mmr_peaks, new_authentication_nodes) = apply_mmr_changes(
             &response.block_header,
             found_relevant_note,
-            current_partial_mmr,
+            current_chain_mmr.partial_mmr_mut(),
             response.mmr_delta,
         )?;
 
-        state_sync_update.block_updates.extend(BlockUpdates::new(
-            vec![(response.block_header, found_relevant_note, new_mmr_peaks)],
-            new_authentication_nodes,
-        ));
+        let mut new_blocks = vec![];
+        if found_relevant_note || response.chain_tip == new_block_num {
+            // Only track relevant blocks or the chain tip
+            new_blocks.push((response.block_header, found_relevant_note, new_mmr_peaks));
+        }
+
+        state_sync_update
+            .block_updates
+            .extend(BlockUpdates::new(new_blocks, new_authentication_nodes));
 
         if response.chain_tip == new_block_num {
             Ok(false)